pyproject.toml
[project]
name = "optra"
version = "0.1.0"
description = "Aggregate and summarize your work history from Slack, Notion, and more"
requires-python = ">=3.11"
license = "MIT"
keywords = ["work", "history", "slack", "notion", "productivity", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "slack-sdk>=3.33.0",
    "notion-client>=2.2.0",
    "anthropic>=0.42.0",
    "sqlmodel>=0.0.22",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]

[project.urls]
Homepage = "https://github.com/RyaninKR/optra"
Repository = "https://github.com/RyaninKR/optra"

[project.scripts]
optra = "optra.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["optra", "server"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["optra"]
